cover MCP tool protocol behavior
