+++ packages/intent-mcp/tests/fixtures/empty/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "intent-mcp-empty-fixture",
+  "private": true
+}
