+++ packages/intent-mcp/tests/fixtures/workspace/package.json
@@ -1,0 +1,4 @@
+{
+  "name": "intent-mcp-workspace-fixture",
+  "private": true
+}
