+++ packages/intent-mcp/tsconfig.json
@@ -4,5 +4,5 @@
     "rootDir": ".",
     "noEmit": true
   },
-  "include": ["src"]
+  "include": ["src", "tests"]
 }
